requirements.txt
joblib==0.14.1
lifelines==0.24.5
matplotlib==3.1.1
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pytorch==1.3.1
scikit-learn==0.20.2
scipy==1.3.2
seaborn==0.9.0
tqdm==4.35.0
